Added about page resposiveness left
