+++ minerva-frontend/src/app/components/image-carousel/styles.css
