+++ minerva-frontend/src/app/globals.css
@@ -25,8 +25,6 @@ body {
 @font-face {
   font-family: "Teko";
   src: url('./fonts/Teko-Medium.ttf') format('truetype');
-  font-family: "Verdana";
-  src: url("./fonts/verdana.ttf") format("truetype");
   font-weight: normal;
   font-style: normal;
 }
